+++ CHANGES.md
@@ -14,12 +14,22 @@
 - [CHANGE] `Sora.Config` 中にあるキャプチャラに関するフィールドを `Sora.CameraConfig` に移動する
     - 修正方法は （TODO: @miosakuma がドキュメントへのリンクに差し替える） を参照して下さい
     - @melpon
+- [UPDATE] SoraClientContext を利用してコードを短くする
+  - @melpon
 - [UPDATE] Sora C++ SDK を `2023.12.0` に上げる
     - @melpon
 - [UPDATE] libwebrtc を `m116.5845.6.1` に上げる
     - @melpon
 - [ADD] 接続中にキャプチャラを切り替える機能を実装
     - @melpon
+- [ADD] デバイスを掴まないようにする `NoVideoDevice`, `NoAudioDevice` を追加
+  - @melpon
+- [ADD] ハードウェアエンコーダを利用するかどうかを設定する `UseHardwareEncoder` を追加
+  - @melpon
+- [ADD] `SelectedSignalingURL` と `ConnectedSignalingURL` プロパティを追加
+  - @melpon
+- [FIX] IosAudioInit を初回接続の場合のみ呼び出すようにすることで、iOS で連続して接続しようとすると落ちることがあったのを修正
+  - @melpon
 
 ## 2023.3.0 (2023-08-08)